requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
spacy==3.7.2
python-docx==1.1.0
